requirements.txt
python-telegram-bot==20.0a2
bs4
requests
